+ Glow/Bloom + New Watermark + Added AntiVoid + Removed a fuck-ton of useless shits + BloxD Helper (not good) + BMCSpeed + BMC Fly + ClientTheme + HUD RECODE (again) + MixinItemRenderer (I FUCKING HATE MIXINS) + poppins_bold.ttf + Step + Mospixel Fly + Mospixel Speed + Fixed Killaura None & Lockview rots + MASSIVE ChestStealer, InvManager, AutoArmor, AutoSort improvements + Verussy Fly + New TargetHUD ~ Renamed Velocity2 to ModeVelo because retarded module sorting no like numbers  ~ Moved AntiBot to combat  + More i probably forgot about!

+++ src/main/java/keystrokesmod/clickgui/components/impl/CategoryComponent.java
@@ -212,7 +212,7 @@ public class CategoryComponent {
 
         if (!this.n4m) {
             int prevY = this.y;
-            this.y = (int) this.moduleY;
+            this.y = this.moduleY;
             if ((this.opened || smoothTimer != null) && !this.modules.isEmpty()) {
                 for (Component c2 : this.modules) {
                     c2.render();
