+++ src/main/java/keystrokesmod/module/impl/combat/AutoClicker.java
@@ -202,7 +202,7 @@ public class AutoClicker extends Module {
 
     public void cookie() {
         double c = Utils.getRandomValue(minCPS, maxCPS, this.rand) + 0.4D * this.rand.nextDouble();
-        long d = (long) ((int) Math.round(1000.0D / c));
+        long d = (int) Math.round(1000.0D / c);
         if (System.currentTimeMillis() > this.k) {
             if (!this.n && this.rand.nextInt(100) >= 85) {
                 this.n = true;
