+++ src/main/java/keystrokesmod/utility/BlockUtils.java
@@ -112,7 +112,7 @@ public class BlockUtils {
             return false;
         } else {
             for(int offset = 0; offset < (int)mc.thePlayer.posY + 2; offset += 2) {
-                AxisAlignedBB bb = mc.thePlayer.getEntityBoundingBox().offset(0.0, (double)(-offset), 0.0);
+                AxisAlignedBB bb = mc.thePlayer.getEntityBoundingBox().offset(0.0, -offset, 0.0);
                 if (!mc.theWorld.getCollidingBoundingBoxes(mc.thePlayer, bb).isEmpty()) {
                     return true;
                 }
